refactor: rename prisma client file
